requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.0
idna==3.4
markdown2==2.4.10
Pygments==2.16.1
python-decouple==3.8
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svgwrite==1.4.3
typing_extensions==4.8.0
urllib3==2.0.7
wavedrom==2.0.3.post3
whitenoise==6.6.0
